Ekstrakcja BandData i BandType do data_types.h
